+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
-name  = Lie_Symmetry_Analysis_Algorithm
-description = Lie symmetry Analysis Algorithm
+name  = symmetries
+description = Lie symmetry analysis algorithm
 long_description = file: README.md
 license = Proprietary
 classifiers = 
@@ -19,4 +19,4 @@ include_package_data = True
 
 [options.packages.find]
 where = src
-include = Lie_Symmetry_Analysis_Algorithm
+include = symmetries
