+++ src/symmetries.egg-info/PKG-INFO
@@ -1,8 +1,10 @@
 Metadata-Version: 2.1
-Name: Lie-Symmetry-Analysis-Algorithm
-Version: 0.1.dev29+g9855ca7.d20220903
-Summary: Lie symmetry Analysis Algorithm
+Name: symmetries
+Version: 0.1.dev30+g4c8ba94.d20220916
+Summary: Lie symmetry analysis algorithm
+Home-page: UNKNOWN
 License: Proprietary
+Platform: UNKNOWN
 Classifier: Programming Language :: Python :: 3
 
 # Lie Symmetry Analysis Agorithm
@@ -15,3 +17,5 @@ The algorithm is able to find the reduced form of the determinant equations. How
 To Run the algorithm the algorithm you first need to install the requirements. It can be done by running the following command:
 
 pip install -r requirements.txt
+
+
