+++ src/symmetries.egg-info/SOURCES.txt
@@ -2,19 +2,23 @@
 README.md
 pyproject.toml
 requirements.txt
+requirements2.txt
 setup.cfg
 setup.py
 Mathematica/IntroToSymmetry.m
 Mathematica/IntroToSymmetry.nb
+Notebooks/Agrupaciones shido.ipynb
 Notebooks/Point_Symmetries_Example_Runs.ipynb
+Notebooks/testing/LieSymmetryAnalysis_Test.ipynb
+Notebooks/testing/functions_draft.ipynb
 data_test/data_algebra/data_compare_derivatives.json
 data_test/data_algebra/data_drop_constants.json
 data_test/data_algebra/data_is_zero.json
 data_test/data_algebra/data_key_ordering.json
 data_test/data_latex/data_dict_to_latex.json
-src/Lie_Symmetry_Analysis_Algorithm.egg-info/PKG-INFO
-src/Lie_Symmetry_Analysis_Algorithm.egg-info/SOURCES.txt
-src/Lie_Symmetry_Analysis_Algorithm.egg-info/dependency_links.txt
-src/Lie_Symmetry_Analysis_Algorithm.egg-info/top_level.txt
+src/symmetries.egg-info/PKG-INFO
+src/symmetries.egg-info/SOURCES.txt
+src/symmetries.egg-info/dependency_links.txt
+src/symmetries.egg-info/top_level.txt
 tests/conftest.py
 tests/test_alebra_operations.py
