+++ src/symmetries.egg-info/dependency_links.txt
