+++ src/symmetries.egg-info/top_level.txt
